+++ src/pages/resultsView/plots/PlotsTabUtils.spec.ts
@@ -22,10 +22,7 @@ import {
     logScalePossible,
 } from './PlotsTabUtils';
 import { Mutation, Sample, Gene } from 'cbioportal-ts-api-client';
-import {
-    AnnotatedNumericGeneMolecularData,
-    AnnotatedMutation,
-} from '../ResultsViewPageStore';
+import { AnnotatedNumericGeneMolecularData } from '../ResultsViewPageStore';
 import { AlterationTypeConstants, DataTypeConstants } from 'shared/constants';
 import { MutationCountBy, AxisMenuSelection } from './PlotsTab';
 import {
@@ -40,6 +37,7 @@ import {
 
 import _ from 'lodash';
 import { GenericAssayTypeConstants } from 'shared/lib/GenericAssayUtils/GenericAssayConfig';
+import { AnnotatedMutation } from 'shared/model/AnnotatedMutation';
 
 describe('PlotsTabUtils', () => {
     describe('makeClinicalAttributeOptions', () => {
