+++ src/shared/components/mutationMapper/MutationMapper.tsx
@@ -43,9 +43,9 @@ import WindowStore from '../window/WindowStore';
 
 import styles from './mutationMapper.module.scss';
 import { ProteinImpactType } from 'cbioportal-frontend-commons';
-import { AnnotatedMutation } from 'pages/resultsView/ResultsViewPageStore';
 import DriverAnnotationProteinImpactTypeBadgeSelector from 'pages/resultsView/mutation/DriverAnnotationProteinImpactTypeBadgeSelector';
 import { Mutation, PtmSource } from 'cbioportal-utils';
+import { AnnotatedMutation } from 'shared/model/AnnotatedMutation';
 
 export interface IMutationMapperProps {
     store: MutationMapperStore;
